[android] Fixed review notes - renamed base fragments https://github.com/mapsme/omim/pull/10273#discussion_r252116996
