+++ android/src/com/mapswithme/maps/auth/BaseAuthFragment.java
@@ -1,15 +1,18 @@
-package com.mapswithme.maps.base;
+package com.mapswithme.maps.auth;
 
-import android.content.Context;
 import android.content.Intent;
 import android.support.annotation.CallSuper;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import com.mapswithme.maps.auth.Authorizer;
-import com.mapswithme.maps.auth.TargetFragmentCallback;
+import com.mapswithme.maps.base.BaseMwmToolbarFragment;
 
-public abstract class NoToolbarBaseMwmAuthFragment extends BaseAsyncOperationFragment
+/**
+ * A base toolbar fragment which is responsible for the <b>authorization flow</b>,
+ * starting from the getting an auth token from a social network and passing it to the core
+ * to get user authorized for the MapsMe server (Passport).
+ */
+public abstract class BaseAuthFragment extends BaseMwmToolbarFragment
     implements Authorizer.Callback, TargetFragmentCallback
 {
   @NonNull
@@ -22,17 +25,17 @@ public abstract class NoToolbarBaseMwmAuthFragment extends BaseAsyncOperationFra
 
   @Override
   @CallSuper
-  public void onAttach(Context context)
+  public void onStart()
   {
-    super.onAttach(context);
+    super.onStart();
     mAuthorizer.attach(this);
   }
 
   @Override
   @CallSuper
-  public void onDestroyView()
+  public void onStop()
   {
-    super.onDestroyView();
+    super.onStop();
     mAuthorizer.detach();
   }
